camera: Set the preferred preview size from the preview sizes table
This branch assumed that there is always a matching video and preview
size for all supported resolutions. However, this is not true in msm8974
for 4k resolutions.

While 4k resolutions can be used with 4k preview sizes, non-4k
resolutions cannot be used with 4k preview sizes. As such, the camera
will report a supported recording size of 4k, but the preview size will
at most be 1080p. The HAL was setting the preferred preview size as the
largest video size which will fail.

Fixes android.hardware.cts.CameraTest#testParameters

Change-Id: I0689671edba0f39db993968443edb5e14bf15f52