fingerprint: Cache the enrolled fingerprints during enrollment or authentication
The HAL was previously unable to return the list of fingerprints during
an enrollment or authentication step. This is mainly due to the fact
that those operations are asynchronous and the backing service is
unavailable to handle that request.

This changeset caches the enrolled fingerprints before entering those
states and returns a cached version whenever the backing service is
busy.

Change-Id: Ia2de0cdbdb28fc64b94dfc34dcf678f09e0c8a67
Ticket: HAM-23