ham: Set a list of fallback ECCs when no sims are available
ro.ril.ecclist is the next last fallback ECC list used when no sim
is present in the phone. Otherwise, it will fallback to a fixed list
that does not contain some of the ECC numbers we need.

As a result calling the chinese ECC 120 when no SIM is present causes
the call to be treated as a regular call instead of an emergency
number.

This patch sets ro.ril.ecclist to a more extensive list than the default
(112,911,000,08,110,118,119,999).

Change-Id: I43b823e034719f79f64fc6e9c864b1b97d9b1443
Ticket: HAM-237