ham: Use the platform double-tap-to-wake
Ham was still using the 12.1 TapToWake hal which doesn't exist anymore,
this patchset sets the tap to wake node path to the correct place so
that the power hal can properly wake up the screen.

Note that it defaults to off, just like in 12.1 because of the added
power usage.

Change-Id: I0ce7158f64fd7c5d6c3bfd751129395fa2cb49b6
Ticket: HAM-1235